+++ library/src/main/java/com/getstream/sdk/chat/adapter/MessageListItem.kt
@@ -8,10 +8,10 @@ import java.util.Date
 import java.util.zip.CRC32
 import java.util.zip.Checksum
 
-sealed class MessageListItem(
-    val isMine: Boolean = false,
-    val messageReadBy: MutableList<ChannelUserRead> = mutableListOf()
-) {
+sealed class MessageListItem {
+    abstract val isMine: Boolean
+    abstract val messageReadBy: MutableList<ChannelUserRead>
+
     fun getStableId(): Long {
         val checksum: Checksum = CRC32()
         val plaintext = when (this) {
@@ -33,15 +33,28 @@ sealed class MessageListItem(
         is ThreadSeparatorItem -> copy()
     }.exhaustive
 
-    data class DateSeparatorItem(val date: Date) : MessageListItem()
+    data class DateSeparatorItem @JvmOverloads constructor(
+        val date: Date,
+        override val isMine: Boolean = false,
+        override val messageReadBy: MutableList<ChannelUserRead> = mutableListOf()
+    ) : MessageListItem()
 
-    data class MessageItem(
+    data class MessageItem @JvmOverloads constructor(
         val message: Message,
         val positions: List<MessageViewHolderFactory.Position> = listOf(),
-        val isMessageMine: Boolean
-    ) : MessageListItem(isMine = isMessageMine)
+        override val isMine: Boolean = false,
+        override val messageReadBy: MutableList<ChannelUserRead> = mutableListOf()
+    ) : MessageListItem()
 
-    data class TypingItem(val users: List<User>) : MessageListItem()
+    data class TypingItem @JvmOverloads constructor(
+        val users: List<User>,
+        override val isMine: Boolean = false,
+        override val messageReadBy: MutableList<ChannelUserRead> = mutableListOf()
+    ) : MessageListItem()
 
-    data class ThreadSeparatorItem(val date: Date = Date()) : MessageListItem()
+    data class ThreadSeparatorItem @JvmOverloads constructor(
+        val date: Date = Date(),
+        override val isMine: Boolean = false,
+        override val messageReadBy: MutableList<ChannelUserRead> = mutableListOf()
+    ) : MessageListItem()
 }