+++ library/src/main/java/com/getstream/sdk/chat/viewmodel/messages/MessageListItemLiveData.java
@@ -94,11 +94,11 @@ public class MessageListItemLiveData extends LiveData<MessageListItemWrapper> {
 
             for (int i = messageItems.size(); i-- > 0; ) {
                 final MessageListItem.MessageItem messageListItem = messageItems.get(i);
-                // skip message owner as reader
                 if (userRead.getUserId().equals(messageListItem.getMessage().getUser().getId())) {
-                    continue;
+                    // last read message is added by this user, so break
+                    break;
                 }
-                if (userRead.getLastRead().after(messageListItem.getMessage().getCreatedAt())) {
+                if (!userRead.getLastRead().before(messageListItem.getMessage().getCreatedAt())) {
                     // set the read state on this entity
                     messageListItem.getMessageReadBy().add(userRead);
                     // we only show it for the last message, so break